requirements.txt
torch>=2.0.0
torchvision>=0.15.0
openslide-python>=1.2.0
Pillow>=9.0.0
opencv-python>=4.7.0
shapely>=2.0.0
tqdm>=4.64.0
numpy>=1.21.0